Remove keyring lazy implementation
